+++ src/core/BUILD.gn
@@ -729,6 +729,7 @@ openthread_core_files = [
   "thread/mle_types.hpp",
   "thread/mlr_manager.cpp",
   "thread/mlr_manager.hpp",
+  "thread/mlr_types.cpp",
   "thread/mlr_types.hpp",
   "thread/neighbor.cpp",
   "thread/neighbor.hpp",
