+++ src/core/CMakeLists.txt
@@ -255,6 +255,7 @@ set(COMMON_SOURCES
     thread/mle_tlvs.cpp
     thread/mle_types.cpp
     thread/mlr_manager.cpp
+    thread/mlr_types.cpp
     thread/neighbor.cpp
     thread/neighbor_table.cpp
     thread/network_data.cpp
